+++ amplitude.js
@@ -118,7 +118,11 @@ var Identify = require('./identify');
 var type = require('./type');
 
 var log = function(s) {
-  console.log('[Amplitude] ' + s);
+  try {
+    console.log('[Amplitude] ' + s);
+  } catch (e) {
+    // console logging not available
+  }
 };
 
 var IDENTIFY_EVENT = '$identify';
@@ -3542,7 +3546,11 @@ var AMP_OP_SET_ONCE = '$setOnce';
 var AMP_OP_UNSET = '$unset';
 
 var log = function(s) {
-  console.log('[Amplitude] ' + s);
+  try {
+    console.log('[Amplitude] ' + s);
+  } catch (e) {
+    // console logging not available
+  }
 };
 
 var Identify = function() {